Add Old Vanish Back
You may ask.. Why bring back old vanish?

Well, some worlds is over 1GB or more or have no storage to turn on spectator mode, and import it back into their realm, plus the old vanish was better, sure, it couldn't fly though walls and stuff but it's better to have this still in the current version of scythe today + in my option it's a way better way to spy on targets who may be hacking or other things.

Otherwise, we should keep the old vanish has a option if they want to use it or not, if not, that's fine, if you want it, then it can happen if MrDiamond64 accepts this pull request.