remove: Remove anti-commandblocks optional module
This optional module should have been removed long ago for several reasons:

1. Performance
Every single tick, each player would have a 26x10x26 area removed of command blocks. A massive operation like that would greatly impact performance

2. Redundant
If the overridecommandblocksenabled module was set to disable the commandblocksenabled gamerule, it would make any and all command blocks inoperable.

3. Unnecessary
No type of cheat can allow for a normal player to place command blocks with a custom command. The only exception to this is if another ForceOP exploit came out, but at that point it would not matter if this module was enabled or not.

Overrall, it is a much better idea to use the overridecommandblocksenabled optional module to impair the ability of command blocks to function instead of rely on the removed module to remove command blocks.