Gametest beta 1.20.70 (#128)
* fix: Add support for MC version 1.20.70

* refactor: Convert anti-gamemode checks to Scripting API

* feat: Improve upon config updater

* refactor: Remove OverrideCommandBlocksEnabled

* fix: Fix ScaffoldE false flags

Lily pads exhibit special behavior where you can place it onto liquid tile blocks. This causes a false flag to occur when there is no valid blocks nearby.

* fix: Fix NoSlow/A false positives

A player can move a boat while they are on top of it. The boat can move at a speed that fits in the NoSlowA speed range and causes a false flag.

* refactor: Improve code quality with custom tags

* refactor: Change msg.sender to msg.player

* fix: remove redundant scoreboard object creations

* fix: Remove references to scoreboard objectives that are no longer needed

* refactor: Update messages sent by functions

* fix: Fix !ui banning the wrong player