refactor: Remove some entity files
These entity files are the same ones in the vanilla game but with is_summonable disabled. This was a precaution set to make sure that a hacker could not summon these entities using Command Block Exploit cheats to wreck havoc on a Realm or server. Since the exploits have been patched for a long time now, these files are redundant.