feat: add Reach/C
ReachB (still needs to be implemented) and ReachC were removed in Scythe close to three years ago in commit 7651c38 as there would be false positives depending on device. 1.21.41 allows the Scripting API to get the player's device so we can now reintroduce this check and make it more sensitive depending on the player's device.

Some additional testing needs to be done with Consoles (which I do not have) to properly finish this check though.