refactor: Don't pass redundant variables into flag function
The 7th parameter for the flag function is the cancel object, which is used to cancel block break, messages, or whatever for easy use. However these checks already cancel the message before the flag function is ran (as there is a 1 tick delay before it is run) making the cancel object redundant.