feat: Migrate BadPackets[1] to be a Scripting API check
I first attempted to migrate BadPackets[1] to a Scripting API check way back in 2022, however the Scripting API had a bug where entering a boat would make a player have absurdly wrong viewing angles. Because of that, the check was reverted back to a function-based check. Refer to #40 for additional info.