feat: Use target selectors in kick command
Originally the /kick command supported target selectors until Minecraft 1.17.10, as exploiters were maliciously using the command with Command Block Exploit. It has now returned in 1.21.60.

The reason why this is so useful is because we no longer have to put the kicked player's name in the command anymore, meaning that usernames that have been spoofed to include invalid characters such as double quotes or blackslashes no longer break commands. This makes the anticheat more resilent against attacks.