refactor: Remove onTick animation
This animation was designed to counteract exploits that could entirely disable the anticheat, back when Scythe was completely dependent on functions. If a malicious user were to be able to change the functioncommandlimit gamerule to 0, with an exploit such as command block exploit for example, none of the Scythe function files would run, effectively disabling the anticheat.

Now that Scythe no longer has any checks that utilize functions anymore, this method to patch disabler exploits is redundant, even more so as exploits to arbitrarily run any commands have been patched. This animation has been running two commands every single tick, for each player online. This means for a Realm with 10 active players, 20 commands were being run per tick just to deal with this particualr exploit.

By removing this animation, we are able to improve Scythe's performance and allow it to handle servers with larger player counts