revert: Go back to using the tick event for InventoryModsB instead of playerInventoryItemChange event
In commit 244140f883b55168ed38b63e3aee5f59c0ecb39b (and subsequently 2a8a96a57fcbd50d65263e5d46bfcba786941d7c), InventoryModsB was changed from checking if a player's cursor item changed while moving to check if an item in the player's inventory changed. This made the check more efficient as Scythe would no longer have to keep track of cursor items every single tick and check if the player moved aswell, and made the check more powerful as it could detect more types of inventory move cheats that changes the players inventory when it should not have.

Unfourtunately there were side-cases which I had not accounted for which could lead to false positives. One of the conditions required for a flag was to check if the old item stack was not previously air. This was to avoid Scythe from flagging players when they picked up an item and the item went into an empty item slot. The problem is that if the player already had the item in their inventory and was under the stack limit, picking up the item would add the item to the pre-exisiting stack. Scythe would see that an item changed in the inventory, however the item count changed and not the slot of the item.

We could just check and see the item typeId is the same, however I realized of another edge case that could result in a false positive. If a server operator or a command block were to change an item in a player's inventory through the '/replaceitem' command, the playerInventoryItemChange event would trigger and result in a false positive due to an item being replaced with another one.

The only way to solve these false positives is to rely on the older version of the InventoryModsB check which relied on player cursor items.