refactor: Use Object.hasOwn for more robust checks with untrusted user input Previously to check if a property existed on an object, we access the property on an object through bracket notation. The problem is that this would check the object's entire prototype chain to see if the property exists. Virtually all Javascript objects include properties such as '__proto__' or 'constructor' which originates from the Object protoype.
For places where untrusted user input was used with this existance check, a user could pass in an argument such as '__proto__' incorrectly pass the existance checks that Scythe uses. This could lead to undefined behavior, however more importantly, if a user were to set their name to one of these property names, they could bypass Scythe's ban system entirely as it would assume their name is apart of the unban queue. Protections to the unban queue against this type of attack have been implemented previously, but a new bypass could potentially exist in the future.

To fix these types of flaws entirely, instead of using bracket notation to check if a property exists on an object, we now use Object.hasOwn(). hasOwn only checks if the property exists in the object itself and not any other objects part of its prototype chain. This means that passing in values such as '__proto__' will no longer pass the property existance checks and incorrectly assume it exists.